VFS: rewrote vfs_traverse
I rewrote the vfs_traverse function so that it would be a little bit
clearer and more commented. Changing up the README for the first time in
a year or so. With this change, of having vfs.c be split up, I am
questioning the naming convention and the existing names of functions.

+++ src/c/include/lib/resource.h
@@ -54,7 +54,7 @@ struct ARC_Resource {
 	/// Driver function group (supplied on init by caller).
 	int dri_group;
 	/// Specific driver function set (supplied on init by caller).
-	int dri_index;
+	uint64_t dri_index;
 	/// Driver functions.
 	struct ARC_DriverDef *driver;
 };
